requirements.txt
amqp==5.1.1
apns2==0.7.2
asgiref==3.7.2
billiard==4.1.0
cachetools==5.3.1
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.2.0
click==8.1.5
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
cryptography==41.0.2
Django==4.2.3
django-autoslug==1.9.9
django-cacheops==7.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-countries==7.5.1
django-google-maps==0.13.0
django-phonenumber-field==7.1.0
django-push-notifications==3.0.0
django-timezone-field==5.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djangorestframework-gis==1.0
funcy==1.18
google-api-core==2.11.1
google-api-python-client==2.93.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
h2==2.6.2
hpack==3.0.0
http-ece==1.1.0
httplib2==0.22.0
hyper==0.7.0
hyperframe==3.2.0
idna==3.4
kombu==5.3.1
msal==1.22.0
oauthlib==3.2.2
phonenumbers==8.13.16
Pillow==10.0.0
prompt-toolkit==3.0.39
protobuf==4.23.4
psycopg2-binary==2.9.6
py-vapid==1.9.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==1.7.1
pyotp==2.8.0
pyparsing==3.1.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
pywebpush==1.14.0
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sentry-sdk==1.28.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
